requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
astunparse==1.6.3
async-timeout==4.0.3
blinker==1.8.2
cachetools==5.5.0
click==8.1.7
datasets==3.0.2
decorator==4.4.2
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
joblib==1.4.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.10.0
networkx==3.2.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
pandas==2.2.2
pillow==10.4.0
proglog==0.1.10
propcache==0.2.0
protobuf==3.20.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
pygame==2.6.1
pywin32==305.1
regex==2024.9.11
requests-oauthlib==2.0.0
rich==13.9.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.17.0
seqeval==1.2.2
setproctitle==1.3.3
smmap==5.0.1
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
tqdm==4.66.5
tzdata==2024.1
wandb==0.18.5
watchdog==5.0.3
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.16.0
